@@ -1,30 +1,32 @@
 name: flutter_plugin_tools
-description: Productivity utils for hosting multiple plugins within one repository.
-publish_to: none
+description: Productivity utils for flutter/plugins and flutter/packages
+repository: https://github.com/flutter/plugins/tree/master/script/tool
+version: 0.1.2
 
 dependencies:
-  args: "^1.4.3"
-  path: "^1.6.1"
-  http: "^0.12.1"
-  async: "^2.0.7"
-  yaml: "^2.1.15"
-  quiver: "^2.0.2"
-  pub_semver: ^1.4.2
-  colorize: ^2.0.0
-  git: ^1.0.0
-  platform: ^2.2.0
-  pubspec_parse: "^0.1.4"
-  test: ^1.6.4
-  meta: ^1.1.7
-  file: ^5.0.10
-  uuid: ^2.0.4
-  http_multi_server: ^2.2.0
-  collection: 1.14.13
+  args: ^2.1.0
+  async: ^2.6.1
+  collection: ^1.15.0
+  colorize: ^3.0.0
+  file: ^6.1.0
+  git: ^2.0.0
+  http: ^0.13.3
+  http_multi_server: ^3.0.1
+  meta: ^1.3.0
+  path: ^1.8.0
+  platform: ^3.0.0
+  pub_semver: ^2.0.0
+  pubspec_parse: ^1.0.0
+  quiver: ^3.0.1
+  test: ^1.17.3
+  uuid: ^3.0.4
+  yaml: ^3.1.0
 
 dev_dependencies:
-  matcher: ^0.12.6
-  mockito: ^4.1.1
-  pedantic: 1.8.0
+  build_runner: ^2.0.3
+  matcher: ^0.12.10
+  mockito: ^5.0.7
+  pedantic: ^1.11.0
 
 environment:
-  sdk: ">=2.3.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
