@@ -2,17 +2,6 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart=2.9
+import 'package:integration_test/integration_test_driver.dart';
 
-import 'dart:convert';
-import 'dart:io';
-import 'package:flutter_driver/flutter_driver.dart';
-
-Future<void> main() async {
-  final FlutterDriver driver = await FlutterDriver.connect();
-  final String data =
-      await driver.requestData(null, timeout: const Duration(minutes: 1));
-  await driver.close();
-  final Map<String, dynamic> result = jsonDecode(data) as Map<String, dynamic>;
-  exit(result['result'] == 'true' ? 0 : 1);
-}
+Future<void> main() => integrationDriver();
