@@ -2,19 +2,17 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-// @dart=2.9
-
 import 'dart:io';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:path_provider_linux/path_provider_linux.dart';
 import 'package:integration_test/integration_test.dart';
+import 'package:path_provider_linux/path_provider_linux.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
 
   testWidgets('getTemporaryDirectory', (WidgetTester tester) async {
     final PathProviderLinux provider = PathProviderLinux();
-    final String result = await provider.getTemporaryPath();
+    final String? result = await provider.getTemporaryPath();
     _verifySampleFile(result, 'temporaryDirectory');
   });
 
@@ -23,26 +21,30 @@ void main() {
       return;
     }
     final PathProviderLinux provider = PathProviderLinux();
-    final String result = await provider.getDownloadsPath();
+    final String? result = await provider.getDownloadsPath();
     _verifySampleFile(result, 'downloadDirectory');
   });
 
   testWidgets('getApplicationDocumentsDirectory', (WidgetTester tester) async {
     final PathProviderLinux provider = PathProviderLinux();
-    final String result = await provider.getApplicationDocumentsPath();
+    final String? result = await provider.getApplicationDocumentsPath();
     _verifySampleFile(result, 'applicationDocuments');
   });
 
   testWidgets('getApplicationSupportDirectory', (WidgetTester tester) async {
     final PathProviderLinux provider = PathProviderLinux();
-    final String result = await provider.getApplicationSupportPath();
+    final String? result = await provider.getApplicationSupportPath();
     _verifySampleFile(result, 'applicationSupport');
   });
 }
 
 /// Verify a file called [name] in [directoryPath] by recreating it with test
 /// contents when necessary.
-void _verifySampleFile(String directoryPath, String name) {
+void _verifySampleFile(String? directoryPath, String name) {
+  expect(directoryPath, isNotNull);
+  if (directoryPath == null) {
+    return;
+  }
   final Directory directory = Directory(directoryPath);
   final File file = File('${directory.path}${Platform.pathSeparator}$name');
 
