@@ -2,6 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+// @dart=2.9
+
 import 'dart:convert';
 
 import 'package:args/command_runner.dart';
@@ -173,8 +175,6 @@ void main() {
                   'Runner',
                   '-destination',
                   'foo_destination',
-                  'CODE_SIGN_IDENTITY=""',
-                  'CODE_SIGNING_REQUIRED=NO',
                   'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
                 ],
                 pluginExampleDirectory2.path),
@@ -231,8 +231,6 @@ void main() {
                   'Runner',
                   '-destination',
                   'id=1E76A0FD-38AC-4537-A989-EA639D7D012A',
-                  'CODE_SIGN_IDENTITY=""',
-                  'CODE_SIGNING_REQUIRED=NO',
                   'GCC_TREAT_WARNINGS_AS_ERRORS=YES',
                 ],
                 pluginExampleDirectory.path),
